HEEDLS-366 - Reformatted Service structure and Admin Session recording
